+++ pkg/vault/vault.go
@@ -3,8 +3,11 @@ package vault
 import (
 	"bytes"
 	"io"
+	"os"
+	"path/filepath"
 
 	"filippo.io/age"
+	"github.com/xaaha/hulak/pkg/utils"
 )
 
 type AgeKey struct {
@@ -42,4 +45,15 @@ func DecryptText(cypherText []byte, identities ...age.Identity) ([]byte, error) 
 		return nil, err
 	}
 	return plaintext, nil
+}
+
+// SetPublicKey writes the public encryption key to .hulak/key.pub in the project root.
+func SetPublicKey(publicEncKey string) error {
+	markerPath, err := utils.GetProjectMarker()
+	if err != nil {
+		return err
+	}
+
+	pubKeyPath := filepath.Join(markerPath, publicKeyFile)
+	return os.WriteFile(pubKeyPath, []byte(publicEncKey), utils.FilePer)
 }
